+++ src/SegundaParte/Pocima.java
@@ -2,7 +2,7 @@ package SegundaParte;
 
 public interface Pocima {
 	
-	public int agregarPocima(Atributo atributo);
+	public double agregarPocima(Atributo atributo);
 	public String getNombre();
 	
 }
