+++ src/SegundaParte/PocimaNumeroEspecifico.java
@@ -2,16 +2,16 @@ package SegundaParte;
 
 
 public class PocimaNumeroEspecifico implements Pocima {
-	private int valor;
+	private double valor;
 	private String nombre;
 	
-	public PocimaNumeroEspecifico(int valor, String nombre) {
+	public PocimaNumeroEspecifico(double valor, String nombre) {
 		this.valor = valor;
 		this.nombre = nombre;
 	}
 	
 	@Override
-	public int agregarPocima(Atributo atributo) {
+	public double agregarPocima(Atributo atributo) {
 		return valor;
 		
 	}
